Improve file/folder processing

+++ run.py
@@ -40,10 +40,27 @@ def parse_arguments():
     return parser.parse_args()
 
 
-def create_temp_directory():
-    temp_dir = "tmp"
+def process_audio_file(file_path, temp_dir, args):
+    call_whisperx(
+        file_path,
+        temp_dir,
+        output_format="srt",
+        speaker_count=args.speaker_count,
+        device_index=args.device_index,
+    )
+
+    if args.format == "lrc":
+        convert_to_lrc(temp_dir, args.output)
+def process_directory(directory_path, args):
+    temp_dir = "tmp" if args.format == "lrc" else args.output
     os.makedirs(temp_dir, exist_ok=True)
-    return temp_dir
+
+    for filename in os.listdir(directory_path):
+        file_path = os.path.join(directory_path, filename)
+        process_audio_file(file_path, temp_dir, args)
+
+    if args.format == "lrc":
+        shutil.rmtree(temp_dir)
 
 
 if __name__ == "__main__":
@@ -52,17 +69,11 @@ if __name__ == "__main__":
     if not os.path.exists(args.input_path):
         raise FileNotFoundError(f"Input path {args.input_path} does not exist.")
 
-    temp_dir = create_temp_directory() if args.format == "lrc" else args.output
-
-    call_whisperx(
-        args.input_path,
-        temp_dir,
-        output_format="srt",
-        speaker_count=args.speaker_count,
-        device_index=args.device_index,
-    )
-
-    # Convert to LRC format if specified, then remove the temporary directory
-    if args.format == "lrc":
-        convert_to_lrc(temp_dir, args.output)
-        shutil.rmtree(temp_dir)
+    if os.path.isdir(args.input_path):
+        process_directory(args.input_path, args)
+    else:
+        temp_dir = "tmp" if args.format == "lrc" else args.output
+        os.makedirs(temp_dir, exist_ok=True)
+        process_audio_file(args.input_path, temp_dir, args)
+        if args.format == "lrc":
+            shutil.rmtree(temp_dir)
